Allow specific non-subcrafts to be piloted inside other crafts.
For example, this fixes bombers being unable to be piloted inside aircraft carriers.

We achieve this by creating a `PlayerSubCraft` for these crafts. This is a nominal subcraft; it should not and does not implement the `SubCraft` interface proper to avoid issues with other plugins (e.g. our fork of Movecraft-Cannons).

The `PlayerSubCraft`'s hitbox is then subtracted from that of its parent craft. This allows the parent craft to move without also moving the `PlayerSubCraft`'s blocks, which would cause the `PlayerSubCraft` to sink.

To allow the `PlayerSubCraft` to be piloted in the first place, we need to relax the conditions under which a craft is considered "already controlled". A `PlayerSubCraft` will not be created if:

  - it would be of the same type of the parent craft (i.e. someone trying to double-pilot a ship)
  - it does not have the `canPlayerPilotInsideAnotherCraft` flag set to `true` (e.g. to prevent people piloting already-piloted crafts as, say, shipyards).